# log1pexp