--- a/src/basicfuns.jl
+++ b/src/basicfuns.jl
@@ -159,20 +159,32 @@ See:
 Note: different than Maechler (2012), also uses bounds specific to Float32 and Float16.
 """
 function log1pexp(x::Real)
-    threshold = _log1pexp_threshold(x)
-    if x > threshold # log1pexp(x) ≈ x for large x
-        return oftype(threshold, x)
+    thresh_lin = _log1pexp_threshold_linear(x)
+    if x > thresh_lin
+        return oftype(thresh_lin, x)
+    elseif x > _log1pexp_threshold_exp(x)
+        return x + exp(-x)
     else
     	return log1p(exp(x))
   	end
 end
 
 # returns a threshold such that log1pexp(x) ≈ x for x > threshold
-@generated function _log1pexp_threshold(::T) where {T<:Real}
+@generated function _log1pexp_threshold_linear(::T) where {T<:Real}
     F = float(T)
-    ϵ = big(eps(F)) / 2
-    thresh = -log(expm1(ϵ))
-    return convert(F, thresh)
+    thresh = -log(expm1(big(eps(F)) / 2))
+    return convert(F, max(thresh, log(big(ℯ) - 1)))
+end
+# tighter thresholds can be obtained numerically. Since this is the fastest branch,
+# we do it here for common types.
+@generated _log1pexp_threshold_linear(::Float64) = 33.3e0
+@generated _log1pexp_threshold_linear(::Float32) = 15.9f0
+
+# returns a threshold such that log1pexp(x) ≈ x + exp(-x) for x > threshold
+@generated function _log1pexp_threshold_exp(::T) where {T<:Real}
+    F = float(T)
+    thresh = -log(big(eps(F))) / 2
+    return convert(F, max(thresh, log(big(ℯ) - 1)))
 end
 
 """