--- a/src/basicfuns.jl
+++ b/src/basicfuns.jl
@@ -184,7 +184,7 @@ where the tolerances of the approximations ≈ are on the order of eps(T)
     ϵ = big(eps(F))
     x0 = log(ϵ / 2)
     x1 = -log(ϵ) / 2
-    x2 = -log(ϵ/2) - log(-log(ϵ/2)) - log(-log(ϵ/2)) / log(ϵ/2) # root of e^-x == x * ϵ/2
+    x2 = -x0 - log(-x0) * (1 + 1 / x0) # root of e^-x == x * ϵ/2
     return (F(x0), F(x1), F(x2))
 end