--- a/src/basicfuns.jl
+++ b/src/basicfuns.jl
@@ -190,7 +190,7 @@ end
 #= For Float64, Float32 we can hard-code the thresholds to (hopefully) reduce compilation
 times. We will use here the exact same thresholds that were used before the function
 _log1pexp_thresholds (which outputs close but not identical thresholds) was introduced,
-to decrease the risk of introducing breaking changes. =#
+or those from Maechler 2012, to decrease the risk of introducing breaking changes. =#
 @inline _log1pexp_thresholds(::Float64) = (-37e0, 18e0, 33e0)
 @inline _log1pexp_thresholds(::Float32) = (-17f0, 9f0, 16f0)