--- a/src/basicfuns.jl
+++ b/src/basicfuns.jl
@@ -202,12 +202,15 @@ should optimize away all computations done here.
 end
 
 #= For Float64, Float32 we can hard-code the thresholds to make absolutely sure they are not
-recompued each time. For Float64 we use the same exact thresholds given by Maechler 2012,
+recompued each time. Also, _log1pexp_thresholds is not completely elided by the
+compiler in Julia 1.0 / 1.6 which LogExpFunctions intends to support.
+For Float64 we use the same exact thresholds given by Maechler 2012,
 since these were the ones used before _log1pexp_thresholds was introduced (which outputs
 close but not identical thresholds), to reduce any (small) risk of breakage.
-For Float32 we use truncated versions of the output of _log1pexp_thresholds.  =#
+For Float32 and Float16 we use truncated versions of the output of _log1pexp_thresholds. =#
 @inline _log1pexp_thresholds(::Float64) = (-37e0, 18e0, 33e0)
-@inline _log1pexp_thresholds(::Float32) = (-17f0, 9f0, 16f0)
+@inline _log1pexp_thresholds(::Float32) = (-17f0, 8f0, 14f0)
+@inline _log1pexp_thresholds(::Float16) = (Float16(-7.6), Float16(3.5), Float16(5.9))
 
 """
 $(SIGNATURES)