Based on
https://cran.r-project.org/web/packages/Rmpfr/vignettes/log1mexp-note.pdf

Close #13