@cossio 2022-03-07
The ChainRules tests fail for `Float32`. But I noticed that `logistic` itself has no tests for `Float32`. Are we sure its implementation is correct?