@cossio 2022-03-08
So now all the thresholds are computed generically. There is a small change for `Float64` in that the original thresholds (currently in master and those in the paper) are: `(-37.0, 18.0, 33.0)`, whereas with the generic implementation in this PR they become: `(-36.7368005696771, 18.021826694558577, 33.231118823529634)`. Note that in the paper they are selected arbitrarily to make them round numbers, so I don't think this is a significant change. Though if we are very picky it could be slightly breaking.

If you do consider this breaking, I can fix this PR to make it fallback to the Maechler bounds (even for Float16 and BigFloat, which is not accurate, I have some examples of this). And then I can make another PR with the generic bound computation which would work for Float16, BigFloat, etc.

Honestly I wouldn't consider the change in bounds from `(-37.0, 18.0, 33.0)` to `(-36.73, 18.02, 33.23)` as significant, but I defer to your better judgement.

Here are updated benchmarks.

This PR:

```
julia> @benchmark log1pexp(x) setup=(x=30randn();) samples=10^7
BenchmarkTools.Trial: 263652 samples with 1000 evaluations.
 Range (min … max):   2.790 ns … 89.320 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     18.866 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   17.326 ns ±  8.482 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▄               ▃               █▁             ▄             
  █▂▁▁▁▁▁▁▁▁▁▁▁█▂▂█▃▂▂▁▁▁▁▁▁▁▁▁▁▁▂██▂▄▃▁▃▁▁▁▁▁▁▂▁█▁▇▁▃▁▃▁▁▃▁▁ ▂
  2.79 ns         Histogram: frequency by time        32.1 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

Current master:

```
julia> @benchmark log1pexp(x) setup=(x=30randn();) samples=10^7
BenchmarkTools.Trial: 219712 samples with 999 evaluations.
 Range (min … max):   9.339 ns … 132.006 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     20.783 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   20.694 ns ±   8.275 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

    █▄                 ▇ ▅            ▁  ▁                      
  ▇▇██▆▃▄▄▅▆▂▁▁▁▂▂▅▆█▂▇█▄█▃▂▁▂▁▄▄▂▅▂▅▂█▅▂█▃▂▂▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▃
  9.34 ns         Histogram: frequency by time         43.2 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```